Move cli into cli project.
