+++ kafka-coroutines-retry/src/main/kotlin/io/provenance/kafka/coroutines/retry/Extensions.kt
@@ -1,0 +1,10 @@
+package io.provenance.kafka.coroutines.retry
+
+import java.nio.ByteBuffer
+
+fun Int.toByteArray() =
+    ByteBuffer.allocate(Int.SIZE_BYTES).putInt(this).array()
+
+fun ByteArray.toInt() =
+    ByteBuffer.wrap(this).int
+
